requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

ansible
setuptools>=11.3
python-etcd
python-dateutil
PyYAML
gevent>=1.0